@@ -2,14 +2,14 @@
 
   <PropertyGroup>
     <OutputType>Exe</OutputType>
-    <TargetFramework>netcoreapp3.1</TargetFramework>
+    <TargetFramework>net6.0</TargetFramework>
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="Microsoft.Extensions.Logging.Console" Version="3.1.0" />
-    <PackageReference Include="Microsoft.Orleans.OrleansTelemetryConsumers.Counters" Version="3.3.0" />
-    <PackageReference Include="Microsoft.Orleans.Server" Version="3.3.0" />
-    <PackageReference Include="OrleansDashboard" Version="3.1.0" />
+    <PackageReference Include="Microsoft.Extensions.Logging.Console" />
+    <PackageReference Include="Microsoft.Orleans.OrleansTelemetryConsumers.Counters" />
+    <PackageReference Include="Microsoft.Orleans.Server" />
+    <PackageReference Include="OrleansDashboard" />
   </ItemGroup>
 
   <ItemGroup>
