@@ -1,9 +1,8 @@
 ﻿using System.Threading.Tasks;
 
-namespace Kritner.Orleans.GettingStarted.GrainInterfaces
+namespace Kritner.Orleans.GettingStarted.GrainInterfaces;
+
+public interface IEmailSender
 {
-    public interface IEmailSender
-    {
-        Task SendEmailAsync(string from, string[] to, string subject, string body);
-    }
+    Task SendEmailAsync(string from, string[] to, string subject, string body);
 }
