@@ -1,10 +1,9 @@
-﻿using Orleans;
-using System.Threading.Tasks;
+﻿using System.Threading.Tasks;
+using Orleans;
 
-namespace Kritner.Orleans.GettingStarted.GrainInterfaces
+namespace Kritner.Orleans.GettingStarted.GrainInterfaces;
+
+public interface IHelloWorld : IGrainWithGuidKey, IGrainInterfaceMarker
 {
-    public interface IHelloWorld : IGrainWithGuidKey, IGrainInterfaceMarker
-    {
-        Task<string> SayHello(string name);
-    }
+    Task<string> SayHello(string name);
 }
