@@ -1,12 +1,11 @@
-﻿using Orleans;
-using System.Threading.Tasks;
+﻿using System.Threading.Tasks;
+using Orleans;
 
-namespace Kritner.Orleans.GettingStarted.GrainInterfaces
+namespace Kritner.Orleans.GettingStarted.GrainInterfaces;
+
+public interface IObservableManager : IGrainWithIntegerKey, IGrainInterfaceMarker
 {
-    public interface IObservableManager : IGrainWithIntegerKey, IGrainInterfaceMarker
-    {
-        Task Subscribe(IObserverSample observer);
-        Task Unsubscribe(IObserverSample observer);
-        Task SendMessageToObservers(string message);
-    }
+    Task Subscribe(IObserverSample observer);
+    Task Unsubscribe(IObserverSample observer);
+    Task SendMessageToObservers(string message);
 }
