@@ -1,10 +1,9 @@
-﻿using Orleans;
-using System;
+﻿using System;
+using Orleans;
 
-namespace Kritner.Orleans.GettingStarted.GrainInterfaces
+namespace Kritner.Orleans.GettingStarted.GrainInterfaces;
+
+public interface IObserverSample : IGrainObserver
 {
-    public interface IObserverSample : IGrainObserver
-    {
-        void ReceiveMessage(string message);
-    }
+    void ReceiveMessage(string message);
 }
