@@ -2,10 +2,9 @@
 using System.Collections.Generic;
 using System.Text;
 
-namespace Kritner.Orleans.GettingStarted.Grains
+namespace Kritner.Orleans.GettingStarted.Grains;
+
+public class Constants
 {
-    public class Constants
-    {
-        public const string OrleansMemoryProvider = "OrleansMemoryProvider";
-    }
+    public const string OrleansMemoryProvider = "OrleansMemoryProvider";
 }
