@@ -1,14 +1,13 @@
-﻿using Kritner.Orleans.GettingStarted.GrainInterfaces;
+﻿using System.Threading.Tasks;
+using Kritner.Orleans.GettingStarted.GrainInterfaces;
 using Orleans;
-using System.Threading.Tasks;
 
-namespace Kritner.Orleans.GettingStarted.Grains
+namespace Kritner.Orleans.GettingStarted.Grains;
+
+public class HelloWorld : Grain, IHelloWorld, IGrainMarker
 {
-    public class HelloWorld : Grain, IHelloWorld, IGrainMarker
+    public Task<string> SayHello(string name)
     {
-        public Task<string> SayHello(string name)
-        {
-            return Task.FromResult($"Hello from grain {this.GetGrainIdentity()}, {name}!");
-        }
+        return Task.FromResult($"Hello from grain {this.GetGrainIdentity()}, {name}!");
     }
 }
