@@ -2,21 +2,20 @@
 using System.Collections.Generic;
 using System.Text;
 
-namespace Kritner.OrleansGettingStarted.Client.Helpers
-{
-    public static class ConsoleHelpers
-    {
-        public static void LineSeparator()
-        {
-            Console.WriteLine($"{Environment.NewLine}-----{Environment.NewLine}");
-        }
+namespace Kritner.OrleansGettingStarted.Client.Helpers;
 
-        public static void ReturnToMenu()
-        {
-            LineSeparator();
-            Console.WriteLine("Press any key to return to menu");
-            Console.ReadKey();
-            Console.Clear();
-        }
+public static class ConsoleHelpers
+{
+    public static void LineSeparator()
+    {
+        Console.WriteLine($"{Environment.NewLine}-----{Environment.NewLine}");
+    }
+
+    public static void ReturnToMenu()
+    {
+        LineSeparator();
+        Console.WriteLine("Press any key to return to menu");
+        Console.ReadKey();
+        Console.Clear();
     }
 }
