@@ -1,14 +1,13 @@
-﻿using Orleans;
-using System;
+﻿using System;
 using System.Collections.Generic;
 using System.Text;
 using System.Threading.Tasks;
+using Orleans;
 
-namespace Kritner.OrleansGettingStarted.Client.OrleansFunctionExamples
+namespace Kritner.OrleansGettingStarted.Client.OrleansFunctionExamples;
+
+public interface IOrleansFunction
 {
-    public interface IOrleansFunction
-    {
-        string Description { get; }
-        Task PerformFunction(IClusterClient clusterClient);
-    }
+    string Description { get; }
+    Task PerformFunction(IClusterClient clusterClient);
 }
