@@ -1,44 +1,43 @@
-﻿using Microsoft.Extensions.Configuration;
+﻿using System;
+using System.IO;
+using Kritner.OrleansGettingStarted.Common.Config;
+using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.FileProviders;
 using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Hosting.Internal;
-using System;
-using System.IO;
-using Kritner.OrleansGettingStarted.Common.Config;
 using Microsoft.Extensions.Options;
 
-namespace Kritner.OrleansGettingStarted.Common.Helpers
+namespace Kritner.OrleansGettingStarted.Common.Helpers;
+
+public static class ConsoleAppConfigurator
 {
-    public static class ConsoleAppConfigurator
+    public static (string env, IConfigurationRoot configurationRoot, OrleansConfig orleansConfig) BootstrapConfigurationRoot()
     {
-        public static (string env, IConfigurationRoot configurationRoot, OrleansConfig orleansConfig) BootstrapConfigurationRoot()
+        var env = GetEnvironmentName();
+        var tempConfigBuilder = new ConfigurationBuilder();
+
+        tempConfigBuilder
+            .AddJsonFile($"appsettings.json", optional: false, reloadOnChange: false)
+            .AddJsonFile($"appsettings.{env}.json", optional: false, reloadOnChange: false);
+
+        var configurationRoot = tempConfigBuilder.Build();
+
+        var serviceCollection = new ServiceCollection();
+        serviceCollection.Configure<OrleansConfig>(configurationRoot.GetSection(nameof(OrleansConfig)));
+        var serviceProvider = serviceCollection.BuildServiceProvider();
+        var orleansConfig = serviceProvider.GetService<IOptions<OrleansConfig>>().Value;
+        return (env, configurationRoot, orleansConfig);
+    }
+
+    public static string GetEnvironmentName()
+    {
+        var env = Environment.GetEnvironmentVariable("ASPNETCORE_ENVIRONMENT");
+        if (string.IsNullOrWhiteSpace(env))
         {
-            var env = GetEnvironmentName();
-            var tempConfigBuilder = new ConfigurationBuilder();
-
-            tempConfigBuilder
-                .AddJsonFile($"appsettings.json", optional: false, reloadOnChange: false)
-                .AddJsonFile($"appsettings.{env}.json", optional: false, reloadOnChange: false);
-
-            var configurationRoot = tempConfigBuilder.Build();
-
-            var serviceCollection = new ServiceCollection();
-            serviceCollection.Configure<OrleansConfig>(configurationRoot.GetSection(nameof(OrleansConfig)));
-            var serviceProvider = serviceCollection.BuildServiceProvider();
-            var orleansConfig = serviceProvider.GetService<IOptions<OrleansConfig>>().Value;
-            return (env, configurationRoot, orleansConfig);
+            throw new Exception("ASPNETCORE_ENVIRONMENT env variable not set.");
         }
-        
-        public static string GetEnvironmentName()
-        {
-            var env = Environment.GetEnvironmentVariable("ASPNETCORE_ENVIRONMENT");
-            if (string.IsNullOrWhiteSpace(env))
-            {
-                throw new Exception("ASPNETCORE_ENVIRONMENT env variable not set.");
-            }
 
-            return env;
-        }
+        return env;
     }
 }
