@@ -4,20 +4,19 @@ using Kritner.Orleans.GettingStarted.Grains;
 using Kritner.Orleans.GettingStarted.Grains.HealthChecks;
 using Microsoft.Extensions.DependencyInjection;
 
-namespace Kritner.OrleansGettingStarted.SiloHost.Helpers
+namespace Kritner.OrleansGettingStarted.SiloHost.Helpers;
+
+/// <summary>
+/// Dependency Injection helper class
+/// </summary>
+public static class DependencyInjectionHelper
 {
     /// <summary>
-    /// Dependency Injection helper class
+    /// Register concretions for DI.
     /// </summary>
-    public static class DependencyInjectionHelper
+    /// <param name="serviceCollection">The service collection in which to register thingers.</param>
+    public static void IocContainerRegistration(IServiceCollection serviceCollection)
     {
-        /// <summary>
-        /// Register concretions for DI.
-        /// </summary>
-        /// <param name="serviceCollection">The service collection in which to register thingers.</param>
-        public static void IocContainerRegistration(IServiceCollection serviceCollection)
-        {
-            serviceCollection.AddSingleton<IEmailSender, FakeEmailSender>();
-        }
+        serviceCollection.AddSingleton<IEmailSender, FakeEmailSender>();
     }
 }
